Enable bash tool sandboxing
Configure filesystem and network isolation for the bash tool, in
exchange for allowing commands to be auto-approved. This is safer than
passing --dangerously-skip-permissions, but still allows a fairly high
level of auto-approval in theory.

Directories used by the sandbox like /tmp/claude must be pre-created,
otherwise it doesn't work. There doesn't appear to be a working way to
add write access to additional directories, so stuff like golang needs
to be told to work in /tmp/claude.